requirements.txt
# --- Core web ---
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
gunicorn==22.0.0

# --- Drivers DB ---
PyMySQL==1.1.1           # MySQL local
psycopg2-binary==2.9.9   # (opcional) Postgres en Render

# --- Data & ML ---
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
scikit-learn==1.4.2
joblib==1.4.2

# --- Visualización / archivos ---
matplotlib==3.8.4
reportlab==4.2.5

# --- Utilidades ---
qrcode==7.4.2
pillow==10.4.0
python-dotenv==1.0.1
